Move capture tech to separate window
